Refactor getItems function to a new utility file and update imports in product listing components
