+++ src/utils/sort-filter.ts
@@ -1,5 +1,4 @@
 import { ItemParams } from "@/models/product-category-collections/item-params.model";
-import ItemApiResponse from "@/models/product-category-collections/item.model";
 import { SortFilterModel } from "@/models/product-category-collections/sort-filter.model";
 import { itemAPI } from "@/services/item.service";
 
@@ -123,22 +122,6 @@ export function getItemParams({ slug }: { slug?: string[] }): ItemParams {
   };
 
   return itemParams;
-}
-
-export async function getItems({
-  params,
-  token,
-}: {
-  params: ItemParams;
-  token?: string;
-}): Promise<ItemApiResponse | null> {
-  try {
-    const response = await itemAPI.getAllItems({ params, token });
-    return response;
-  } catch (error) {
-    console.error(error);
-    return null;
-  }
 }
 
 export async function getSortFilterOptions(): Promise<{
